+++ eemeter/caltrack/usage_per_day.py
@@ -1,4 +1,4 @@
-from collections import Counter
+from collections import Counter, namedtuple
 import traceback
 
 import numpy as np
@@ -6,30 +6,21 @@ import pandas as pd
 import pytz
 import statsmodels.formula.api as smf
 
-from .api import (
-    CandidateModel,
-    DataSufficiency,
-    EEMeterWarning,
-    ModelResults,
-    ModelPrediction,
-)
-from .exceptions import MissingModelParameterError, UnrecognizedModelTypeError
-from .features import (
-    compute_time_features,
-    compute_temperature_features,
-    compute_temperature_bin_features,
-    compute_occupancy_feature,
-    merge_features,
-)
-from .metrics import ModelMetrics
-from .segmentation import SegmentModel, fit_segmented_model
-from .transform import day_counts, overwrite_partial_rows_with_nan
+from ..exceptions import MissingModelParameterError, UnrecognizedModelTypeError
+from ..features import compute_temperature_features
+from ..metrics import ModelMetrics
+from ..transform import day_counts, overwrite_partial_rows_with_nan
+from ..warnings import EEMeterWarning
 
 
 __all__ = (
-    "caltrack_method",
+    "CalTRACKUsagePerDayCandidateModel",
+    "CalTRACKUsagePerDayModelResults",
+    "DataSufficiency",
+    "ModelPrediction",
+    "fit_caltrack_usage_per_day_model",
     "caltrack_sufficiency_criteria",
-    "caltrack_predict",
+    "caltrack_usage_per_day_predict",
     "plot_caltrack_candidate",
     "get_too_few_non_zero_degree_day_warning",
     "get_total_degree_day_too_low_warning",
@@ -43,41 +34,368 @@ __all__ = (
     "get_hdd_only_candidate_models",
     "get_cdd_hdd_candidate_models",
     "select_best_candidate",
-    "caltrack_hourly_fit_feature_processor",
-    "caltrack_hourly_prediction_feature_processor",
-    "fit_caltrack_hourly_model_segment",
-    "fit_caltrack_hourly_model",
 )
 
 
-def _candidate_model_factory(
-    model_type,
-    formula,
-    status,
-    warnings=None,
-    model_params=None,
-    model=None,
-    result=None,
-    r_squared_adj=None,
-    use_predict_func=True,
-):
-    if use_predict_func:
-        predict_func = caltrack_predict
-    else:
-        predict_func = None
+ModelPrediction = namedtuple("ModelPrediction", ["result", "design_matrix", "warnings"])
 
-    return CandidateModel(
-        model_type=model_type,
-        formula=formula,
-        status=status,
-        warnings=warnings,
-        predict_func=predict_func,
-        plot_func=plot_caltrack_candidate,
-        model_params=model_params,
-        model=model,
-        result=result,
-        r_squared_adj=r_squared_adj,
-    )
+
+def _noneify(value):
+    if value is None:
+        return None
+    return None if np.isnan(value) else value
+
+
+class CalTRACKUsagePerDayModelResults(object):
+    """ Contains information about the chosen model.
+
+    Attributes
+    ----------
+    status : :any:`str`
+        A string indicating the status of this result. Possible statuses:
+
+        - ``'NO DATA'``: No baseline data was available.
+        - ``'NO MODEL'``: No candidate models qualified.
+        - ``'SUCCESS'``: A qualified candidate model was chosen.
+
+    method_name : :any:`str`
+        The name of the method used to fit the baseline model.
+    model : :any:`eemeter.CandidateModel` or :any:`None`
+        The selected candidate model, if any.
+    r_squared_adj : :any:`float`
+        The adjusted r-squared of the selected model.
+    candidates : :any:`list` of :any:`eemeter.CandidateModel`
+        A list of any model candidates encountered during the model
+        selection and fitting process.
+    warnings : :any:`list` of :any:`eemeter.EEMeterWarning`
+        A list of any warnings reported during the model selection and fitting
+        process.
+    metadata : :any:`dict`
+        An arbitrary dictionary of metadata to be associated with this result.
+        This can be used, for example, to tag the results with attributes like
+        an ID::
+
+            {
+                'id': 'METER_12345678',
+            }
+
+    settings : :any:`dict`
+        A dictionary of settings used by the method.
+    totals_metrics : :any:`ModelMetrics`
+        A ModelMetrics object, if one is calculated and associated with this
+        model. (This initializes to None.) The ModelMetrics object contains
+        model fit information and descriptive statistics about the underlying data,
+        with that data expressed as period totals.
+    avgs_metrics : :any:`ModelMetrics`
+        A ModelMetrics object, if one is calculated and associated with this
+        model. (This initializes to None.) The ModelMetrics object contains
+        model fit information and descriptive statistics about the underlying data,
+        with that data expressed as daily averages.
+    """
+
+    def __init__(
+        self,
+        status,
+        method_name,
+        model=None,
+        r_squared_adj=None,
+        candidates=None,
+        warnings=None,
+        metadata=None,
+        settings=None,
+    ):
+        self.status = status  # NO DATA | NO MODEL | SUCCESS
+        self.method_name = method_name
+        self.model = model
+        self.r_squared_adj = r_squared_adj
+
+        if candidates is None:
+            candidates = []
+        self.candidates = candidates
+
+        if warnings is None:
+            warnings = []
+        self.warnings = warnings
+
+        if metadata is None:
+            metadata = {}
+        self.metadata = metadata
+
+        if settings is None:
+            settings = {}
+        self.settings = settings
+
+        self.totals_metrics = None
+        self.avgs_metrics = None
+
+    def __repr__(self):
+        return (
+            "CalTRACKUsagePerDayModelResults(status='{}', method_name='{}',"
+            " r_squared_adj={})".format(
+                self.status, self.method_name, self.r_squared_adj
+            )
+        )
+
+    def json(self, with_candidates=False):
+        """ Return a JSON-serializable representation of this result.
+
+        The output of this function can be converted to a serialized string
+        with :any:`json.dumps`.
+        """
+
+        def _json_or_none(obj):
+            return None if obj is None else obj.json()
+
+        data = {
+            "status": self.status,
+            "method_name": self.method_name,
+            "model": _json_or_none(self.model),
+            "r_squared_adj": _noneify(self.r_squared_adj),
+            "warnings": [w.json() for w in self.warnings],
+            "metadata": self.metadata,
+            "settings": self.settings,
+            "totals_metrics": _json_or_none(self.totals_metrics),
+            "avgs_metrics": _json_or_none(self.avgs_metrics),
+            "candidates": None,
+        }
+        if with_candidates:
+            data["candidates"] = [candidate.json() for candidate in self.candidates]
+        return data
+
+    def plot(
+        self,
+        ax=None,
+        title=None,
+        figsize=None,
+        with_candidates=False,
+        candidate_alpha=None,
+        temp_range=None,
+    ):
+        """ Plot a model fit.
+
+        Parameters
+        ----------
+        ax : :any:`matplotlib.axes.Axes`, optional
+            Existing axes to plot on.
+        title : :any:`str`, optional
+            Chart title.
+        figsize : :any:`tuple`, optional
+            (width, height) of chart.
+        with_candidates : :any:`bool`
+            If True, also plot candidate models.
+        candidate_alpha : :any:`float` between 0 and 1
+            Transparency at which to plot candidate models. 0 fully transparent,
+            1 fully opaque.
+
+        Returns
+        -------
+        ax : :any:`matplotlib.axes.Axes`
+            Matplotlib axes.
+        """
+        try:
+            import matplotlib.pyplot as plt
+        except ImportError:  # pragma: no cover
+            raise ImportError("matplotlib is required for plotting.")
+
+        if figsize is None:
+            figsize = (10, 4)
+
+        if ax is None:
+            fig, ax = plt.subplots(figsize=figsize)
+
+        if temp_range is None:
+            temp_range = (20, 90)
+
+        if with_candidates:
+            for candidate in self.candidates:
+                candidate.plot(ax=ax, temp_range=temp_range, alpha=candidate_alpha)
+        self.model.plot(ax=ax, best=True, temp_range=temp_range)
+
+        if title is not None:
+            ax.set_title(title)
+
+        return ax
+
+
+class CalTRACKUsagePerDayCandidateModel(object):
+    """ Contains information about a candidate model.
+
+    Attributes
+    ----------
+    model_type : :any:`str`
+        The type of model, e..g., :code:`'hdd_only'`.
+    formula : :any:`str`
+        The R-style formula for the design matrix of this model, e.g., :code:`'meter_value ~ hdd_65'`.
+    status : :any:`str`
+        A string indicating the status of this model. Possible statuses:
+
+        - ``'NOT ATTEMPTED'``: Candidate model not fitted due to an issue
+          encountered in data before attempt.
+        - ``'ERROR'``: A fatal error occurred during model fit process.
+        - ``'DISQUALIFIED'``: The candidate model fit was disqualified
+          from the model selection process because of a decision made after
+          candidate model fit completed, e.g., a bad fit, or a parameter out
+          of acceptable range.
+        - ``'QUALIFIED'``: The candidate model fit is acceptable and can be
+          considered during model selection.
+    model_params : :any:`dict`, default :any:`None`
+        A flat dictionary of model parameters which must be serializable
+        using the :any:`json.dumps` method.
+    model : :any:`object`
+        The raw model (if any) used in fitting. Not serialized.
+    result : :any:`object`
+        The raw modeling result (if any) returned by the `model`. Not serialized.
+    r_squared_adj : :any:`float`
+        The adjusted r-squared of the candidate model.
+    warnings : :any:`list` of :any:`eemeter.EEMeterWarning`
+        A list of any warnings reported during creation of the candidate model.
+    """
+
+    def __init__(
+        self,
+        model_type,
+        formula,
+        status,
+        model_params=None,
+        model=None,
+        result=None,
+        r_squared_adj=None,
+        warnings=None,
+    ):
+        self.model_type = model_type
+        self.formula = formula
+        self.status = status  # NOT ATTEMPTED | ERROR | QUALIFIED | DISQUALIFIED
+        self.model = model
+        self.result = result
+        self.r_squared_adj = r_squared_adj
+
+        if model_params is None:
+            model_params = {}
+        self.model_params = model_params
+
+        if warnings is None:
+            warnings = []
+        self.warnings = warnings
+
+    def __repr__(self):
+        return (
+            "CalTRACKUsagePerDayCandidateModel(model_type='{}', formula='{}', status='{}',"
+            " r_squared_adj={})".format(
+                self.model_type,
+                self.formula,
+                self.status,
+                round(self.r_squared_adj, 3)
+                if self.r_squared_adj is not None
+                else None,
+            )
+        )
+
+    def json(self):
+        """ Return a JSON-serializable representation of this result.
+
+        The output of this function can be converted to a serialized string
+        with :any:`json.dumps`.
+        """
+        return {
+            "model_type": self.model_type,
+            "formula": self.formula,
+            "status": self.status,
+            "model_params": self.model_params,
+            "r_squared_adj": _noneify(self.r_squared_adj),
+            "warnings": [w.json() for w in self.warnings],
+        }
+
+    def predict(
+        self,
+        temperature_data,
+        prediction_index,
+        degree_day_method,
+        with_disaggregated=False,
+        with_design_matrix=False,
+    ):
+        return caltrack_usage_per_day_predict(
+            self.model_type,
+            self.model_params,
+            temperature_data,
+            prediction_index,
+            degree_day_method,
+            with_disaggregated=with_disaggregated,
+            with_design_matrix=with_design_matrix,
+        )
+
+    def plot(
+        candidate,
+        best=False,
+        ax=None,
+        title=None,
+        figsize=None,
+        temp_range=None,
+        alpha=None,
+        **kwargs
+    ):
+        return plot_caltrack_candidate(
+            candidate,
+            best=best,
+            ax=ax,
+            title=title,
+            figsize=figsize,
+            temp_range=temp_range,
+            alpha=alpha,
+            **kwargs
+        )
+
+
+class DataSufficiency(object):
+    """ Contains the result of a data sufficiency check.
+
+    Attributes
+    ----------
+    status : :any:`str`
+        A string indicating the status of this result. Possible statuses:
+
+        - ``'NO DATA'``: No baseline data was available.
+        - ``'FAIL'``: Data did not meet criteria.
+        - ``'PASS'``: Data met criteria.
+    criteria_name : :any:`str`
+        The name of the criteria method used to check for baseline data sufficiency.
+    warnings : :any:`list` of :any:`eemeter.EEMeterWarning`
+        A list of any warnings reported during the check for baseline data sufficiency.
+    settings : :any:`dict`
+        A dictionary of settings (keyword arguments) used.
+    """
+
+    def __init__(self, status, criteria_name, warnings=None, settings=None):
+        self.status = status  # NO DATA | FAIL | PASS
+        self.criteria_name = criteria_name
+
+        if warnings is None:
+            warnings = []
+        self.warnings = warnings
+
+        if settings is None:
+            settings = {}
+        self.settings = settings
+
+    def __repr__(self):
+        return (
+            "DataSufficiency("
+            "status='{status}', criteria_name='{criteria_name}')".format(
+                status=self.status, criteria_name=self.criteria_name
+            )
+        )
+
+    def json(self):
+        """ Return a JSON-serializable representation of this result.
+
+        The output of this function can be converted to a serialized string
+        with :any:`json.dumps`.
+        """
+        return {
+            "status": self.status,
+            "criteria_name": self.criteria_name,
+            "warnings": [w.json() for w in self.warnings],
+            "settings": self.settings,
+        }
 
 
 def _get_parameter_or_raise(model_type, model_params, param):
@@ -216,7 +534,7 @@ def _caltrack_predict_design_matrix(
         return base_load + heating_load + cooling_load
 
 
-def caltrack_predict(
+def caltrack_usage_per_day_predict(
     model_type,
     model_params,
     temperature_data,
@@ -579,7 +897,9 @@ def get_fit_failed_candidate_model(model_type, formula):
             data={"traceback": traceback.format_exc()},
         )
     ]
-    return _candidate_model_factory(model_type, formula, "ERROR", warnings)
+    return CalTRACKUsagePerDayCandidateModel(
+        model_type=model_type, formula=formula, status="ERROR", warnings=warnings
+    )
 
 
 def get_intercept_only_candidate_models(data, weights_col):
@@ -631,10 +951,10 @@ def get_intercept_only_candidate_models(data, weights_col):
         status = "QUALIFIED"
 
     return [
-        _candidate_model_factory(
-            model_type,
-            formula,
-            status,
+        CalTRACKUsagePerDayCandidateModel(
+            model_type=model_type,
+            formula=formula,
+            status=status,
             warnings=model_warnings,
             model_params=model_params,
             model=model,
@@ -695,12 +1015,11 @@ def get_single_cdd_only_candidate_model(
     )
 
     if len(degree_day_warnings) > 0:
-        return _candidate_model_factory(
-            model_type,
-            formula,
-            "NOT ATTEMPTED",
+        return CalTRACKUsagePerDayCandidateModel(
+            model_type=model_type,
+            formula=formula,
+            status="NOT ATTEMPTED",
             warnings=degree_day_warnings,
-            use_predict_func=False,
         )
 
     if weights_col is None:
@@ -746,10 +1065,10 @@ def get_single_cdd_only_candidate_model(
     else:
         status = "QUALIFIED"
 
-    return _candidate_model_factory(
-        model_type,
-        formula,
-        status,
+    return CalTRACKUsagePerDayCandidateModel(
+        model_type=model_type,
+        formula=formula,
+        status=status,
         warnings=model_warnings,
         model_params=model_params,
         model=model,
@@ -851,12 +1170,11 @@ def get_single_hdd_only_candidate_model(
     )
 
     if len(degree_day_warnings) > 0:
-        return _candidate_model_factory(
-            model_type,
-            formula,
-            "NOT ATTEMPTED",
+        return CalTRACKUsagePerDayCandidateModel(
+            model_type=model_type,
+            formula=formula,
+            status="NOT ATTEMPTED",
             warnings=degree_day_warnings,
-            use_predict_func=False,
         )
 
     if weights_col is None:
@@ -902,10 +1220,10 @@ def get_single_hdd_only_candidate_model(
     else:
         status = "QUALIFIED"
 
-    return _candidate_model_factory(
-        model_type,
-        formula,
-        status,
+    return CalTRACKUsagePerDayCandidateModel(
+        model_type=model_type,
+        formula=formula,
+        status=status,
         warnings=model_warnings,
         model_params=model_params,
         model=model,
@@ -1047,12 +1365,8 @@ def get_single_cdd_hdd_candidate_model(
     )
 
     if len(degree_day_warnings) > 0:
-        return _candidate_model_factory(
-            model_type,
-            formula,
-            "NOT ATTEMPTED",
-            warnings=degree_day_warnings,
-            use_predict_func=False,
+        return CalTRACKUsagePerDayCandidateModel(
+            model_type, formula, "NOT ATTEMPTED", warnings=degree_day_warnings
         )
 
     if weights_col is None:
@@ -1110,10 +1424,10 @@ def get_single_cdd_hdd_candidate_model(
     else:
         status = "QUALIFIED"
 
-    return _candidate_model_factory(
-        model_type,
-        formula,
-        status,
+    return CalTRACKUsagePerDayCandidateModel(
+        model_type=model_type,
+        formula=formula,
+        status=status,
         warnings=model_warnings,
         model_params=model_params,
         model=model,
@@ -1237,7 +1551,7 @@ def select_best_candidate(candidate_models):
     return best_candidate, []
 
 
-def caltrack_method(
+def fit_caltrack_usage_per_day_model(
     data,
     fit_cdd=True,
     use_billing_presets=False,
@@ -1253,7 +1567,8 @@ def caltrack_method(
     fit_hdd_only=True,
     fit_cdd_hdd=True,
 ):
-    """ CalTRACK method.
+    """ CalTRACK daily and billing methods using a usage-per-day modeling
+    strategy.
 
     Parameters
     ----------
@@ -1300,8 +1615,8 @@ def caltrack_method(
 
     Returns
     -------
-    model_results : :any:`eemeter.ModelResults`
-        Results of running CalTRACK daily method. See :any:`eemeter.ModelResults`
+    model_results : :any:`eemeter.CalTRACKUsagePerDayModelResults`
+        Results of running CalTRACK daily method. See :any:`eemeter.CalTRACKUsagePerDayModelResults`
         for more details.
     """
     if use_billing_presets:
@@ -1314,12 +1629,12 @@ def caltrack_method(
     data = overwrite_partial_rows_with_nan(data)
 
     if data.empty:
-        return ModelResults(
+        return CalTRACKUsagePerDayModelResults(
             status="NO DATA",
-            method_name="caltrack_method",
+            method_name="caltrack_usage_per_day",
             warnings=[
                 EEMeterWarning(
-                    qualified_name="eemeter.caltrack_method.no_data",
+                    qualified_name="eemeter.caltrack_usage_per_day.no_data",
                     description=("No data available. Cannot fit model."),
                     data={},
                 )
@@ -1385,9 +1700,9 @@ def caltrack_method(
         status = "SUCCESS"
         r_squared_adj = best_candidate.r_squared_adj
 
-    model_result = ModelResults(
+    model_result = CalTRACKUsagePerDayModelResults(
         status=status,
-        method_name="caltrack_method",
+        method_name="caltrack_usage_per_day",
         model=best_candidate,
         candidates=candidates,
         r_squared_adj=r_squared_adj,
@@ -1766,118 +2081,3 @@ def plot_caltrack_candidate(
         ax.set_title(title)
 
     return ax
-
-
-def caltrack_hourly_fit_feature_processor(
-    segment_name, segmented_data, occupancy_lookup, temperature_bins
-):
-    # get occupied feature
-    hour_of_week = segmented_data.hour_of_week
-    occupancy = occupancy_lookup[segment_name]
-    occupancy_feature = compute_occupancy_feature(hour_of_week, occupancy)
-
-    # get temperature bin features
-    temperatures = segmented_data.temperature_mean
-    bin_endpoints_list = (
-        temperature_bins[segment_name].index[temperature_bins[segment_name]].tolist()
-    )
-    # TODO(philngo): combine with compute_temperature_features
-    temperature_bin_features = compute_temperature_bin_features(
-        segmented_data.temperature_mean, bin_endpoints_list
-    )
-
-    # combine features
-    return merge_features(
-        [
-            segmented_data[["meter_value", "hour_of_week"]],
-            occupancy_feature,
-            temperature_bin_features,
-            segmented_data.weight,
-        ]
-    )
-
-
-def caltrack_hourly_prediction_feature_processor(
-    segment_name, segmented_data, occupancy_lookup, temperature_bins
-):
-    # hour of week feature
-    hour_of_week_feature = compute_time_features(
-        segmented_data.index,
-        hour_of_week=True,
-        day_of_week=False,
-        hour_of_day=False,
-    )
-
-    # occupancy feature
-    occupancy = occupancy_lookup[segment_name]
-    occupancy_feature = compute_occupancy_feature(
-        hour_of_week_feature.hour_of_week, occupancy
-    )
-
-    # get temperature bin features
-    temperatures = segmented_data
-    bin_endpoints_list = (
-        temperature_bins[segment_name].index[temperature_bins[segment_name]].tolist()
-    )
-    # TODO(philngo): combine with compute_temperature_features
-    temperature_bin_features = compute_temperature_bin_features(
-        segmented_data.temperature_mean, bin_endpoints_list
-    )
-
-    # combine features
-    return merge_features(
-        [
-            hour_of_week_feature,
-            occupancy_feature,
-            temperature_bin_features,
-            segmented_data.weight,
-        ]
-    )
-
-
-def fit_caltrack_hourly_model_segment(segment_name, segment_data):
-    def _get_hourly_model_formula(data):
-        bin_occupancy_interactions = "".join(
-            [" + {}:occupancy".format(c) for c in data.columns if "bin" in c]
-        )
-        return "meter_value ~ C(hour_of_week) - 1{}".format(bin_occupancy_interactions)
-
-    formula = _get_hourly_model_formula(segment_data)
-    model = smf.wls(formula=formula, data=segment_data, weights=segment_data.weight)
-    model_params = {coeff: value for coeff, value in model.fit().params.items()}
-    warnings = []
-    return SegmentModel(
-        segment_name=segment_name,
-        model=model,
-        formula=formula,
-        model_params=model_params,
-        warnings=warnings,
-    )
-
-
-def fit_caltrack_hourly_model(segmented_design_matrices, occupancy_lookup, temperature_bins):
-    segmented_model = fit_segmented_model(
-        segmented_design_matrices,
-        fit_caltrack_hourly_model_segment,
-        prediction_segment_type='one_month',
-        prediction_segment_name_mapping={
-            'jan': 'dec-jan-feb-weighted',
-            'feb': 'jan-feb-mar-weighted',
-            'mar': 'feb-mar-apr-weighted',
-            'apr': 'mar-apr-may-weighted',
-            'may': 'apr-may-jun-weighted',
-            'jun': 'may-jun-jul-weighted',
-            'jul': 'jun-jul-aug-weighted',
-            'aug': 'jul-aug-sep-weighted',
-            'sep': 'aug-sep-oct-weighted',
-            'oct': 'sep-oct-nov-weighted',
-            'nov': 'oct-nov-dec-weighted',
-            'dec': 'nov-dec-jan-weighted',
-        },
-        prediction_feature_processor=caltrack_hourly_prediction_feature_processor,
-        prediction_feature_processor_kwargs={
-            'occupancy_lookup': occupancy_lookup,
-            'temperature_bins': temperature_bins,
-        },
-    )
-    return segmented_model
