+++ tests/test_visualization.py
@@ -4,14 +4,11 @@ import pytest
 
 matplotlib.use("Agg")
 
-from eemeter import (
-    CandidateModel,
-    ModelResults,
-    plot_energy_signature,
-    plot_time_series,
-    plot_caltrack_candidate,
-    caltrack_predict,
+from eemeter.caltrack.usage_per_day import (
+    CalTRACKUsagePerDayCandidateModel,
+    CalTRACKUsagePerDayModelResults,
 )
+from eemeter.visualization import plot_energy_signature, plot_time_series
 
 
 def test_plot_time_series(il_electricity_cdd_hdd_daily):
@@ -34,12 +31,10 @@ def test_plot_energy_signature(il_electricity_cdd_hdd_daily):
 
 
 def test_plot_caltrack_candidate_qualified():
-    candidate_model = CandidateModel(
+    candidate_model = CalTRACKUsagePerDayCandidateModel(
         model_type="intercept_only",
         formula="formula",
         status="QUALIFIED",
-        predict_func=caltrack_predict,
-        plot_func=plot_caltrack_candidate,
         model_params={"intercept": 1},
     )
     ax = candidate_model.plot(candidate_model, title="title")
@@ -49,12 +44,10 @@ def test_plot_caltrack_candidate_qualified():
 
 
 def test_plot_caltrack_candidate_disqualified():
-    candidate_model = CandidateModel(
+    candidate_model = CalTRACKUsagePerDayCandidateModel(
         model_type="intercept_only",
         formula="formula",
         status="DISQUALIFIED",
-        predict_func=caltrack_predict,
-        plot_func=plot_caltrack_candidate,
         model_params={"intercept": 1},
     )
     ax = candidate_model.plot()
@@ -63,12 +56,10 @@ def test_plot_caltrack_candidate_disqualified():
 
 
 def test_plot_caltrack_candidate_with_range():
-    candidate_model = CandidateModel(
+    candidate_model = CalTRACKUsagePerDayCandidateModel(
         model_type="intercept_only",
         formula="formula",
         status="QUALIFIED",
-        predict_func=caltrack_predict,
-        plot_func=plot_caltrack_candidate,
         model_params={"intercept": 1},
     )
     ax = candidate_model.plot(temp_range=(10, 20))
@@ -77,12 +68,10 @@ def test_plot_caltrack_candidate_with_range():
 
 
 def test_plot_caltrack_candidate_best():
-    candidate_model = CandidateModel(
+    candidate_model = CalTRACKUsagePerDayCandidateModel(
         model_type="intercept_only",
         formula="formula",
         status="QUALIFIED",
-        predict_func=caltrack_predict,
-        plot_func=plot_caltrack_candidate,
         model_params={"intercept": 1},
     )
     ax = candidate_model.plot(best=True)
@@ -91,11 +80,10 @@ def test_plot_caltrack_candidate_best():
 
 
 def test_plot_caltrack_candidate_error():
-    candidate_model = CandidateModel(
+    candidate_model = CalTRACKUsagePerDayCandidateModel(
         model_type="intercept_only",
         formula="formula",
         status="ERROR",
-        plot_func=plot_caltrack_candidate,  # no predict func
         model_params={"intercept": 1},
     )
     ax = candidate_model.plot()
@@ -103,12 +91,10 @@ def test_plot_caltrack_candidate_error():
 
 
 def test_plot_caltrack_candidate_cdd_hdd_model():
-    candidate_model = CandidateModel(
+    candidate_model = CalTRACKUsagePerDayCandidateModel(
         model_type="cdd_hdd",
         formula="formula",
         status="QUALIFIED",
-        predict_func=caltrack_predict,
-        plot_func=plot_caltrack_candidate,
         model_params={
             "beta_hdd": 1,
             "beta_cdd": 1,
@@ -123,15 +109,13 @@ def test_plot_caltrack_candidate_cdd_hdd_model():
 
 
 def test_plot_model_results(il_electricity_cdd_hdd_daily):
-    candidate_model = CandidateModel(
+    candidate_model = CalTRACKUsagePerDayCandidateModel(
         model_type="intercept_only",
         formula="formula",
         status="QUALIFIED",
-        predict_func=caltrack_predict,
-        plot_func=plot_caltrack_candidate,
         model_params={"intercept": 1},
     )
-    model_results = ModelResults(
+    model_results = CalTRACKUsagePerDayModelResults(
         status="status",
         method_name="method_name",
         model=candidate_model,
